[HUDI-1029] In inline compaction mode, previously failed compactions needs to be retried before new compactions
  - Prevents failed compactions from causing issues with future commits
  - Need to add tests
